backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.1
email-validator>=2.0.0
pytest>=8.0.0
pytest-asyncio>=1.0.0
httpx>=0.28.0
minio>=7.2.0
boto3>=1.42.51
greenlet>=3.0.0